ingestion_service/requirements.txt
fastapi
uvicorn[standard]
boto3
psycopg2-binary
python-multipart
requests
